% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_check}
\alias{api_check}
\title{A general check for the API and \code{legisTaiwan}.}
\usage{
api_check(start_date = start_date, end_date = end_date)
}
\arguments{
\item{start_date}{start_date is inherited from global env.}

\item{end_date}{end_date is inherited from global env.}
}
\description{
A general check for the API and \code{legisTaiwan}.
}
