% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Checking the date}
\usage{
check_date(roc_date)
}
\arguments{
\item{roc_date}{date format in Taiwan ROC calendar (e.g., "1050531") as
a string vector}
}
\value{
date format in A.D. format
}
\description{
Checking the date
}
\examples{
check_date("1050531")
}
