% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date2}
\alias{check_date2}
\title{Transforming the date in Taiwan ROC calendar to A.D. in POSIXct}
\usage{
check_date2(roc_date)
}
\arguments{
\item{roc_date}{Date format in Taiwan ROC calendar (e.g., "105/05/31") as a
string vector}
}
\value{
date in POSIXct
}
\description{
Transforming the date in Taiwan ROC calendar to A.D. in POSIXct
}
\details{
\code{check_date} transforms ROC date to a date in POSIXct, e.g. "105/05/31" to "2016-05-31".
}
\examples{
x <- check_date2("105/05/31")

}
