% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_internet}
\alias{check_internet}
\title{A check for internet connectivity.}
\usage{
check_internet(x = curl::has_internet())
}
\arguments{
\item{x}{The default value is \code{curl::has_internet()}, which activate the
internet connectivity check.}
}
\description{
A check for internet connectivity.
}
