% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\encoding{UTF-8}
\name{get_bills}
\alias{get_bills}
\title{The Records of the Bills}
\usage{
get_bills(start_date = NULL, end_date = NULL, proposer = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{numeric. Must be formatted in the ROC Taiwan calendar, e.g., 1090101.}

\item{end_date}{numeric. Must be formatted in the ROC Taiwan calendar, e.g., 1090102.}

\item{proposer}{The default value is NULL, indicating that bills proposed by all legislators
are included between the start and end dates.}

\item{verbose}{logical. Specifies whether \code{get_bills} should print out
detailed output when retrieving the data. The default value is TRUE.}
}
\value{
A list, which contains:
\describe{
\item{\code{title}}{Records of cross-caucus sessions}
\item{\code{query_time}}{Query timestamp}
\item{\code{retrieved_number}}{Number of observations retrieved}
\item{\code{meeting_unit}}{Meeting unit}
\item{\code{start_date_ad}}{Start date in POSIXct format}
\item{\code{end_date_ad}}{End date in POSIXct format}
\item{\code{start_date}}{Start date in the ROC Taiwan calendar}
\item{\code{url}}{URL of the retrieved JSON data}
\item{\code{variable_names}}{Variable names of the tibble dataframe}
\item{\code{manual_info}}{Official manual. See \url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=153} or use get_variable_info("get_bills")}
\item{\code{data}}{A tibble dataframe with the following variables:
\describe{\item{\code{term}}{Session number}
\item{\code{sessionPeriod}}{Session period}
\item{\code{sessionTimes}}{Session count}
\item{\code{meetingTimes}}{Proposal date}
\item{\code{billName}}{Bill name}
\item{\code{billProposer}}{Primary proposer}
\item{\code{billCosignatory}}{Co-signatories of the bill}
\item{\code{billStatus}}{Status of the bill}
\item{\code{date_ad}}{Date in the Gregorian calendar}
}
}
}
}
\description{
The Records of the Bills
}
\details{
The \code{get_bills} function returns a list that contains \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info}, and \code{data}.
}
\note{
To retrieve the user manual and more details about the data frame, use \code{get_variable_info("get_bills")}.
Further checks are required as the user manual seems to be inconsistent with the actual data.
}
\examples{
\dontrun{
## Query bill records by a date range in the Taiwan ROC calendar format
get_bills(start_date = 1060120, end_date = 1070310, verbose = TRUE)

## Query bill records by a date range and a specific legislator
get_bills(start_date = 1060120, end_date = 1070310,  proposer = "Kung Wen-chi")

## Query bill records by a date range and multiple legislators
get_bills(start_date = 1060120, end_date = 1060510,  proposer = "Kung Wen-chi&Cheng Tien-tsai")
}

}
\seealso{
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=153}
}
\author{
David Liao (davidycliao@gmail.com)
}
