% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\encoding{UTF-8}
\name{get_bills_2}
\alias{get_bills_2}
\title{The Records of Legislation and the Executives Proposals}
\usage{
get_bills_2(term = 8, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{A numeric or NULL value. Data is available from the 8th term onwards.
Default is set to 8.}

\item{session_period}{An integer, numeric, or NULL. Valid options for the session are:
1, 2, 3, 4, 5, 6, 7, and 8. Default is set to NULL.
\code{review_session_info()} provides available session periods based on the Minguo (Taiwan) calendar.}

\item{verbose}{Default value is TRUE. Displays details of the retrieved data, including the number, URL, and computing time.}
}
\value{
A list containing:
\item{\code{title}}{Records of questions answered by the executives}
\item{\code{query_time}}{Query time}
\item{\code{retrieved_number}}{Number of observations}
\item{\code{retrieved_term}}{Retrieved term}
\item{\code{url}}{Retrieved JSON URL}
\item{\code{variable_names}}{Variables of the tibble dataframe}
\item{\code{manual_info}}{Official manual: \url{https://data.ly.gov.tw/getds.action?id=20} or use \code{get_variable_info("get_bills_2")}}
\item{\code{data}}{A tibble dataframe with variables such as:
\describe{
\item{\code{term}}{Legislative term number}
\item{\code{sessionPeriod}}{Session period}
\item{\code{sessionTimes}}{Session times}
\item{\code{meetingTimes}}{Extraordinary session times}
\item{\code{billNo}}{Bill number}
\item{\code{billName}}{Bill title}
\item{\code{billOrg}}{Proposing unit/legislator}
\item{\code{billProposer}}{Primary proposer}
\item{\code{billCosignatory}}{Co-signed proposal}
\item{\code{billStatus}}{Bill status}
\item{\code{pdfUrl}}{PDF download link for related documents}
\item{\code{docUrl}}{DOC download link for related documents}
\item{\code{selectTerm}}{Filtering criteria based on term}
}
}
}
\description{
The Records of Legislation and the Executives Proposals
}
\details{
The \code{get_bills_2} function produces a list, which includes \code{query_time},
\code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names}, \code{manual_info}, and \code{data}.
For the user manual and more information about the dataframe, use \code{get_variable_info("get_bills_2")}.
}
\note{
For more details about the dataframe's variables, use \code{get_variable_info("get_bills_2")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=20}.
Category: Provides bill-proposal records for legislators and the executive
(available from the 8th term, 1st session, onwards).
}
\examples{
\dontrun{
## Query the executives' responses by term and session period.
get_bills_2(term = 8, session_period = 1)
}

}
\seealso{
\code{get_variable_info("get_bills_2")},\code{review_session_info()}
}
\author{
David Liao (davidycliao@gmail.com)
}
