% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\encoding{UTF-8}
\name{get_caucus_meetings}
\alias{get_caucus_meetings}
\title{The Meeting Records of Cross-caucus Session}
\usage{
get_caucus_meetings(start_date = NULL, end_date = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{character Must be formatted in Minguo (ROC) calendar with three
forward slashes between year, month and day, e.g. "106/10/20".}

\item{end_date}{character Must be formatted in Minguo (ROC) calendar with three
forward slashes between year, month and day, e.g. "109/01/10".}

\item{verbose}{logical, indicates whether \code{get_caucus_meetings} should print out
detailed output when retrieving the data.}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the official manual, \url{https://data.ly.gov.tw/getds.action?id=8}; or use get_variable_info("get_caucus_meetings")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{comYear}}{Volume}
\item{\code{comVolume}}{Issue}
\item{\code{comBookId}}{Book ID}
\item{\code{term}}{Term number}
\item{\code{sessionPeriod}}{Session period}
\item{\code{meetingTimes}}{Extraordinary session times}
\item{\code{meetingDate}}{Meeting date (Minguo calendar)}
\item{\code{meetingName}}{Meeting name}
\item{\code{subject}}{Subject}
\item{\code{pageStart}}{Starting page}
\item{\code{pageEnd}}{Ending page}
\item{\code{docUrl}}{File download location}
\item{\code{htmlUrl}}{HTML URL}
\item{\code{selectTerm}}{Term/session filtering criteria}
}
}
}
}
\description{
Retrieves cross-caucus negotiation meeting records from the Legislative Yuan's
V1 API.
}
\details{
\code{get_caucus_meetings} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data.}
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_caucus_meetings")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=8}.
Category: Provides cross-caucus negotiation records from the gazette
(available from the 8th term, 1st session, onwards)
}
\examples{
## query the meeting records of cross-caucus session using a period of
## the dates in Taiwan ROC calender format with forward slash (/).
get_caucus_meetings(start_date = "106/10/20", end_date = "107/03/10")

}
\seealso{
\code{get_variable_info("get_caucus_meetings")}  Regarding Minguo calendar, please see \url{https://en.wikipedia.org/wiki/Republic_of_China_calendar}.
}
\author{
Yen-Chieh Liao (davidycliao@gmail.com)
}
