% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\encoding{UTF-8}
\name{get_executive_response}
\alias{get_executive_response}
\title{The Records of Response to the Questions by the Executives}
\usage{
get_executive_response(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{integer, numeric or NULL. The default is NULL. The data is only
available from 8th term.}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to NULL.
\code{review_session_info()} generates each session period  available option period
in Minguo (Taiwan) calendar.}

\item{verbose}{logical, indicates whether \code{get_executive_response} should
print out detailed output when retrieving the data. The default is set to TRUE}
}
\value{
list contains: \describe{
\item{\code{title}}{the records of the questions answered by the executives}
\item{\code{query_time}}{the queried time}
\item{\code{retrieved_number}}{the total number of observations}
\item{\code{retrieved_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{term}}{Term number}
\item{\code{sessionPeriod}}{Session period}
\item{\code{sessionTimes}}{Session times}
\item{\code{meetingTimes}}{Extraordinary session times}
\item{\code{eyNumber}}{Executive Yuan document reference number}
\item{\code{lyNumber}}{Legislative Yuan document reference number}
\item{\code{subject}}{Subject}
\item{\code{content}}{Content}
\item{\code{docUrl}}{Subject}
\item{\code{item}}{File download location}
\item{\code{item}}{File download location}
\item{\code{selectTerm}}{Term/session filtering criteria}
}
}
}
}
\description{
Provides access to the records of parliamentary questions through
the Legislative Yuan's V1 API interface.
}
\details{
\strong{\code{get_executive_response}} produces a list, which contains \code{title},
\code{query_time}, \code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names},
\code{manual_info} and \code{data}. To retrieve the user manual and more information, please
use \code{get_variable_info("get_executive_response")}.

#'@note To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_executive_response")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=2}.
Category: Provides records of the Executive Yuan's responses to legislators'
questions in the gazette (available from the 8th term, 1st session, onwards).
}
\examples{
\dontrun{
## query the Executives' answered response by term and the session period.
term8 <- get_executive_response(term = 8, session_period = 1)
term8
}
}
\seealso{
\code{get_variable_info("get_executive_response")}, \code{review_session_info()}
}
\author{
Yen-Chieh Liao (davidycliao@gmail.com)
}
