% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_infos}
\alias{get_infos}
\title{Check function information on the website of Taiwan Legislative API}
\usage{
get_infos(x)
}
\arguments{
\item{x}{The parameter should be \code{get_parlquestions}, \code{get_legislators},
\code{get_executive_response}, \code{get_bills}, or \code{get_meetings}}
}
\description{
Check function information on the website of Taiwan Legislative API
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=6}
\url{https://data.ly.gov.tw/getds.action?id=16}
\url{https://data.ly.gov.tw/getds.action?id=2}
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=153}
}
