% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{get_legislators}
\alias{get_legislators}
\title{The Legislator' Demographic Information and Background}
\usage{
get_legislators(term = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or NULL The data is available from the 2nd term.}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data. The default is set to TRUE.}
}
\value{
list contains: \describe{
\item{\code{query_time}}{the queried time}
\item{\code{queried_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the official manual from \url{https://data.ly.gov.tw/getds.action?id=16}, or use legisTaiwan::get_variable_info("get_legislators")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{term}}{Term number}
\item{\code{name}}{Legislator's name (Chinese)}
\item{\code{ename}}{Legislator's name (English)}
\item{\code{sex}}{Gender}
\item{\code{party}}{Political party affiliation}
\item{\code{partyGroup}}{Party group/caucus}
\item{\code{committee}}{Committee assignment}
\item{\code{onboardDate}}{Onboard date (Gregorian year)}
\item{\code{degree}}{Education background}
\item{\code{experience}}{Work experience}
\item{\code{picPath}}{Photo URL/path}
\item{\code{leaveFlag}}{Leave date (Gregorian year)}
\item{\code{leaveReason}}{Reason for leaving}
}
}
}
}
\description{
Provides access to legislators' basic demographic and background information through
the Legislative Yuan's V1 API interface.
}
\details{
This function retrieves comprehensive data about legislators including their
personal information, political background, and demographic details. The data
is available starting from the 2nd legislative term.

\code{get_legislators} produces a list, which contains  \code{query_time},
\code{queried_term}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}.
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_legislators")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=16}.
Provides legislators' basic information; the earliest available data goes
back to the 2nd term.
}
\examples{
\dontrun{
# Get data for the 9th term
legislators_data <- get_legislators(term = 9)
legislators_data
}

}
\seealso{
\code{get_variable_info("get_legislators")}, \code{review_session_info()}
}
\author{
Yen-Chieh Liao (davidycliao@gmail.com)
}
