% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\encoding{UTF-8}
\name{get_meetings}
\alias{get_meetings}
\title{The Spoken Meeting Records}
\usage{
get_meetings(
  start_date = NULL,
  end_date = NULL,
  meeting_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{numeric Must be formatted in Minguo (Taiwan) calendar, e.g. 1090101.}

\item{end_date}{numeric Must be formatted in Minguo (Taiwan) calendar, e.g. 1090102.}

\item{meeting_unit}{NULL The default is NULL, which includes all meeting types
between the starting date and the ending date.}

\item{verbose}{logical, indicates whether \code{get_meetings} should print out
detailed output when retrieving the data.}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the spoken meeting records }
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of the observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual, \url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}; or use get_variable_info("get_meetings")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{smeeting_date}}{Meeting date}
\item{\code{meeting_status}}{Meeting status}
\item{\code{meeting_name}}{Meeting name}
\item{\code{meeting_content}}{Meeting subject}
\item{\code{speechers}}{List of speaking legislators}
\item{\code{meeting_unit}}{Organizing unit}
\item{\code{date_ad}}{Gregorian year}
}
}
}
}
\description{
Provides access to legislators' spoken meeting records through
the Legislative Yuan's V1 API interface.
}
\details{
\code{get_meetings} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}.
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_meetings")} or visit
the API manual at \url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}.
The data appears inconsistent and needs verification; the earliest available
date for spoken meeting records is unclear and needs checking.
}
\examples{
\dontrun{
## query meeting records by a period of the dates in Minguo (Taiwan) calendar
get_meetings(start_date = "1050120", end_date = "1050210")

## query meeting records by a period of the dates in Minguo (Taiwan) calendar format
## and a meeting; meeting_unit is a Chinese committee name, e.g. "Interior Committee"
get_meetings(start_date = 1060120, end_date = 1070310, meeting_unit = "Interior Committee")
}

}
\seealso{
\code{get_variable_info("get_meetings")}

Regarding Minguo calendar, please see \url{https://en.wikipedia.org/wiki/Republic_of_China_calendar}.
}
\author{
Yen-Chieh Liao (davidycliao@gmail.com)
}
