% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\encoding{UTF-8}
\name{get_parlquestions}
\alias{get_parlquestions}
\title{The Records of Parliamentary Questions}
\usage{
get_parlquestions(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or NULL. The default is set to NULL.}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to 8
\code{review_session_info()} generates each session period available option period
in Minguo (Taiwan) calendar.}

\item{verbose}{logical, indicates whether \code{get_parlquestions} should print out
detailed output when retrieving the data. The default is TRUE.}
}
\value{
A list containing:
\describe{
\item{\code{title}}{the records of parliamentary questions}
\item{\code{query_time}}{the queried time}
\item{\code{retrieved_number}}{the total number of observations}
\item{\code{retrieved_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual from \url{https://data.ly.gov.tw/getds.action?id=6}, or use get_variable_info("get_parlquestions")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{
\item{\code{term}}{Term number}
\item{\code{sessionPeriod}}{Session period}
\item{\code{sessionTimes}}{Session times}
\item{\code{item}}{Item}
\item{\code{selectTerm}}{Term/session filtering criteria}
}
}
}
}
\description{
Provides access to the records of parliamentary questions through
the Legislative Yuan's V1 API interface.
}
\details{
\code{get_parlquestions} produces a list, which contains \code{title},
\code{query_time}, \code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names},
\code{manual_info}, and \code{data}.
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_parlquestions")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=6}.
Category: Provides records of legislators' questions from the meeting agenda
(available from the 8th term, 1st session, onwards).
}
\examples{
\dontrun{
term10 <- get_parlquestions(term = 10, session_period = 1)
term10
}

}
\seealso{
\code{get_variable_info("get_parlquestions")}
}
\author{
Yen-Chieh Liao (davidycliao@gmail.com)
}
