% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{havingIP}
\alias{havingIP}
\title{A simple way to check IP for connectivity}
\usage{
havingIP()
}
\description{
A simple way to check IP for connectivity
}
\seealso{
\url{https://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r?noredirect=1&lq=1}
}
