% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_online}
\alias{is_online}
\title{A simple way to check for the website connectivity}
\usage{
is_online(site = "https://data.ly.gov.tw/index.action")
}
\arguments{
\item{site}{https://data.ly.gov.tw/index.action}
}
\description{
A simple way to check for the website connectivity
}
\seealso{
\url{https://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r?noredirect=1&lq=1}
}
