% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{review_session_info}
\alias{review_session_info}
\title{Check Session Periods in Each Year}
\usage{
review_session_info(term)
}
\arguments{
\item{term}{numeric The term number (1-11) of Legislative Yuan}
}
\value{
A dataframe containing session information
}
\description{
Examines session periods in the Taiwan Minguo calendar
}
\details{
Check Session Periods in Each Year (Minguo Calendar)

The review_session_info function produces a dataframe, displaying each session
period in year formatted in Minguo calendar. This implementation uses
system curl command with --insecure option to bypass SSL/TLS issues.
}
\examples{
\dontrun{
# Show the session information for the 7th Legislative Yuan term periods in ROC calendar year
review_session_info(7)
}
}
\seealso{
Regarding Minguo calendar, please see \url{https://en.wikipedia.org/wiki/Republic_of_China_calendar}.
}
\author{
David Liao (davidycliao@gmail.com)
}
