% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformed_date}
\alias{transformed_date}
\title{A function for transforming date from Taiwan ROC calendar into A.D. format}
\usage{
transformed_date(rocdate)
}
\arguments{
\item{rocdate}{Date format in Taiwan ROC calendar (e.g., "105/05/31") as a string vector}
}
\value{
date
}
\description{
A function for transforming date from Taiwan ROC calendar into A.D. format
}
\examples{
transformed_date("105/05/31")
}
