% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformed_date_bill}
\alias{transformed_date_bill}
\title{Transforming the date in Taiwan ROC calendar to A.D. format for get_bill()}
\usage{
transformed_date_bill(roc_date)
}
\arguments{
\item{roc_date}{date format in Taiwan ROC calendar (e.g., "1050531") as a
string vector}
}
\value{
date format in A.D. format
}
\description{
Transforming the date in Taiwan ROC calendar to A.D. format for get_bill()
}
\examples{
transformed_date_bill("1050531")
}
