% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformed_date_meeting}
\alias{transformed_date_meeting}
\title{Transforming meeting date in Taiwan ROC calendar to A.D. format}
\usage{
transformed_date_meeting(rocdate)
}
\arguments{
\item{rocdate}{Date format in Taiwan ROC calendar (e.g., "105/05/31") as a string vector}
}
\value{
date
}
\description{
Transforming meeting date in Taiwan ROC calendar to A.D. format
}
\examples{
transformed_date_meeting("105/05/31")
}
