% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_dates_format}
\alias{validate_dates_format}
\title{Validate Date Formats}
\usage{
validate_dates_format(start_date, end_date)
}
\arguments{
\item{start_date}{A string representing the start date. Expected format is numeric, e.g., "1090101".}

\item{end_date}{A string representing the end date. Expected format is numeric, e.g., "1090101".}
}
\value{
NULL. If the dates are not in the expected format, an error is thrown.
}
\description{
This function checks if the provided start and end dates are in the correct numeric format.
}
\examples{
# This should throw an error:
\dontrun{
validate_dates_format("10901", "1100101")
}
}
\keyword{internal}
