% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{website_availability}
\alias{website_availability}
\title{A check for the website availability and the connection.}
\usage{
website_availability(site = "https://data.ly.gov.tw/index.action")
}
\arguments{
\item{site}{https://data.ly.gov.tw/index.action}
}
\description{
A check for the website availability and the connection.
}
\seealso{
\code{check_internet()},
}
