% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{website_availability2}
\alias{website_availability2}
\title{A check for the website availability and the connection.}
\usage{
website_availability2(
  site = "https://npl.ly.gov.tw/do/www/appDate?status=0&expire=02&startYear=0n"
)
}
\arguments{
\item{site}{https://npl.ly.gov.tw/do/www/appDate?status=0&expire=02&startYear=0}
}
\description{
A check for the website availability and the connection.
}
\seealso{
\code{check_internet()}, \code{website_availability()}
}
