% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Transforming Minguo (Taiwan) Calendar to A.D. Calendar III}
\usage{
check_date(roc_date)
}
\arguments{
\item{roc_date}{date format in Minguo (Taiwan) Calendar (e.g., "1050531") in a
character vector}
}
\value{
date in POSIXct
}
\description{
\code{transformed_date_meeting()} transforms Minguo (Taiwan) Calendar
to A.D. format in POSIXct for \code{get_bill()}, e.g. "1050531" to "2016-05-31".
}
\keyword{internal}
