% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date2}
\alias{check_date2}
\title{Transforming Minguo (Taiwan) Calendar to A.D. Calendar IIII}
\usage{
check_date2(roc_date)
}
\arguments{
\item{roc_date}{Date format in Minguo (Taiwan) calendar (e.g., "105/05/31") as a
string vector}
}
\value{
date in POSIXct
}
\description{
\code{transformed_date_meeting()} transforms Minguo (Taiwan) Calendar
to A.D. format in POSIXct for \code{get_bill()}, e.g. "1050531" to "2016-05-31".
}
\details{
\code{check_date} transforms ROC date to a date in POSIXct, e.g. "105/05/31" to "2016-05-31".
}
\examples{
\dontrun{
check_date2("105/05/31")
}

}
\keyword{internal}
