% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{get_bills}
\alias{get_bills}
\title{The Records of the Bills 法律提案}
\usage{
get_bills(start_date = NULL, end_date = NULL, proposer = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{numeric Must be formatted in ROC Taiwan calendar, e.g. 1090101.}

\item{end_date}{numeric Must be formatted in ROC Taiwan calendar, e.g. 1090102.}

\item{proposer}{The default value is NULL, which means all bill proposed by all legislators
are included between the starting date and the ending date.}

\item{verbose}{logical, indicates whether \code{get_bills} should print out
detailed output when retrieving the data. The default value is TRUE.}
}
\value{
list, which contains: s\describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\code{term:屆別},
\code{sessionPeriod:會期},
\code{sessionTimes:會次},
\code{meetingTimes:提案日期},
\code{billName:提案名稱},
\code{billProposer:主提案人},
\code{billCosignatory:連署提案},
\verb{billStatus：}, and
\code{date_ad:西元年}}
}
}
\description{
The Records of the Bills 法律提案
}
\details{
\code{get_bills} produces a list, which contains \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}. To retrieve the user
manual and more information about the data frame, please use
\code{legisTaiwan::get_variable_info("get_bills")}. 法律提案
}
\examples{
## query bill records by a period of the dates in Taiwan ROC calender format
## 輸入「中華民國民年」下載立法委員提案資料
get_bills(start_date = 1060120, end_date = 1070310, verbose = FALSE)

## query bill records by a period of the dates in Taiwan ROC calender format
## and a specific legislator
## 輸入「中華民國民年」與「指定立法委員」下載立法委員提案資料
get_bills(start_date = 1060120, end_date = 1070310,  proposer = "孔文吉")

## query bill records by a period of the dates in Taiwan ROC calender format
## and multiple legislators
## 輸入「中華民國民年」與「指定多個立法委員」下載立法委員提案資料
get_bills(start_date = 1060120, end_date = 1060510,  proposer = "孔文吉&鄭天財")
}
\seealso{
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=153}
}
