% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{get_bills_2}
\alias{get_bills_2}
\title{Retrieving the records of legislators and the government (executives) proposals
提供委員及政府之議案提案資訊。(自第8屆第1會期起)}
\usage{
get_bills_2(term = 8, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or NULL The default value is 8
參數必須為數值，資料從自第8屆第1會期起。}

\item{session_period}{numeric or NULL. Available options for the session periods
is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is 8 參數必須為數值。}

\item{verbose}{The default value is TRUE, displaying the description of data
retrieved in number, url and computing time.}
}
\value{
list list, which contains: \describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\code{term},
\code{sessionPeriod},
\code{sessionTimes},
\code{meetingTimes},
\code{billNo},
\code{billName},
\code{billOrg},
\code{billProposer},
\code{billCosignatory},
\code{billStatus},
\code{pdfUrl},
\code{docUrl} and
\code{selectTerm}}
}
}
\description{
Retrieving the records of legislators and the government (executives) proposals
提供委員及政府之議案提案資訊。(自第8屆第1會期起)
}
\details{
\code{get_bills_2} produces a list, which contains \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}.
}
\examples{
## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「質詢事項 (行政院答復部分)」
get_bills_2(term = 8, session_period = 1)

}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=1}
}
