% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_caucus_meetings}
\alias{get_caucus_meetings}
\title{Retrieving the meeting records of cross-caucus session 下載「黨團協商」資料}
\usage{
get_caucus_meetings(start_date = NULL, end_date = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{Requesting meeting records starting from the date.
A double represents a date in ROC Taiwan format.
If a double is used, it should specify as Taiwan
calendar format, e.g. 109/01/10.}

\item{end_date}{Requesting meeting records ending from the date.
A double represents a date in ROC Taiwan format. If a double is used,
it should specify as Taiwan calendar format, e.g. 109/01/20.}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data.}
}
\value{
An object of the list, which contains query_time, retrieved_number,
start_date_ad, end_date_ad, start_date, end_date, url, variable_names,
manual_info and data
}
\description{
Retrieving the meeting records of cross-caucus session 下載「黨團協商」資料
}
\examples{
## query the meeting records of cross-caucus session using a period of
## the dates in Taiwan ROC calender format with forward slash (/).
## 輸入「中華民國民年」下載「黨團協商」，輸入時間請依照該格式 "106/10/20"
## ，需有「正斜線」做隔開。
get_caucus_meetings(start_date = "106/10/20", end_date = "107/03/10")

}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=8}
}
