% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_caucus_meetings}
\alias{get_caucus_meetings}
\title{Retrieving the meeting records of cross-caucus session
議事類:提供公報之黨團協商資訊。(自第8屆第1會期起)}
\usage{
get_caucus_meetings(start_date = NULL, end_date = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{character Must be formatted in ROC Taiwan calendar with three
forward slashes between year, month and day, e.g. "106/10/20".}

\item{end_date}{character Must be formatted in ROC Taiwan calendar with three
forward slashes between year, month and day, e.g. "109/01/10".}

\item{verbose}{logical, indicates whether \code{get_caucus_meetings} should print out
detailed output when retrieving the data.}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\code{comYear: 卷},
\code{comVolume: 期},
\code{comBookId: 冊別},
\code{term: 屆別},
\code{sessionPeriod: 會期},
\code{sessionTimes: 會次},
\code{meetingTimes: 臨時會會次},
\code{meetingDate: 會議日期(民國年)},
\code{meetingName: 會議名稱},
\code{subject: 案由},
\code{pageStart: 起始頁},
\code{pageEnd: 結束頁}
\code{docUrl: 檔案下載位置 },
\code{htmlUrl: html網址}, and
\code{selectTerm: 屆別期別篩選條件}}
}
}
\description{
Retrieving the meeting records of cross-caucus session
議事類:提供公報之黨團協商資訊。(自第8屆第1會期起)
}
\details{
\code{get_caucus_meetings} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data.}
}
\examples{
## query the meeting records of cross-caucus session using a period of
## the dates in Taiwan ROC calender format with forward slash (/).
## 輸入「中華民國民年」下載「黨團協商」，輸入時間請依照該格式 "106/10/20"，
## 需有「正斜線」做隔開。
get_caucus_meetings(start_date = "106/10/20", end_date = "107/03/10")

}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=8}
}
