% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_committee_record}
\alias{get_committee_record}
\title{The Records of Reviewed Items in the Committees 委員會會議審查之議案項目}
\usage{
get_committee_record(term = 8, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or null. The data is only available from 8th term.
The default is set to 8. 參數必須為數值。資料從自第8屆起，預設值為8。}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to NULL. 參數必須為數值。
\code{review_session_info()} generates each session period  available option period
in Minguo (Taiwan) calendar.}

\item{verbose}{logical, indicates whether \code{get_executive_response} should
print out detailed output when retrieving the data. The default is TRUE}
}
\value{
list contains: \describe{
\item{\code{title}}{the records of the questions answered by the executives}
\item{\code{query_time}}{the queried time}
\item{\code{retrieved_number}}{the total number of observations}
\item{\code{retrieved_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the official manual, \url{https://data.ly.gov.tw/getds.action?id=46}; or use legisTaiwan::get_variable_info("get_committee_record")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{term}}{屆別}
\item{\code{sessionPeriod}}{會期}
\item{\code{meetingNo}}{會議編號}
\item{\code{billNo}}{議案編號}
\item{\code{selectTerm}}{屆別期別篩選條件}
}
}
}
}
\description{
The Records of Reviewed Items in the Committees 委員會會議審查之議案項目
}
\details{
\code{get_committee_record} produces a list, which contains \code{title},
\code{query_time}, \code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names},
\code{manual_info} and \code{data}.
}
\note{
To retrieve the user manual and more information
about variable of the data frame, please use \code{legisTaiwan::get_variable_info("get_committee_record")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=46}.
議事類: 提供委員會會議審查之議案項目 (自第8屆第1會期起)。
}
\examples{
## query the committee record by term and the session period.
## 輸入「立委屆期」與「會期」下載「委員會審議之議案」
get_committee_record(term = 8, session_period = 1)

}
\seealso{
\code{get_variable_info("get_committee_record")}, \code{review_session_info()}
}
\author{
David Yen-Chieh Liao
}
