% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\name{get_executive_response}
\alias{get_executive_response}
\title{The Records of the Questions Answered by the Executives 公報質詢事項行政院答復資訊}
\usage{
get_executive_response(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{integer, numeric or null. The default is NULL. The data is only
available from 8th term. 參數必須為數值，資料從自第8屆起。}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is NULL. 參數必須為數值。}

\item{verbose}{logical, indicates whether \code{get_executive_response} should
print out detailed output when retrieving the data. The default is TRUE}
}
\value{
list contains: \describe{
\item{\code{title}}{the records of the questions answered by the executives}
\item{\code{query_time}}{the queried time}
\item{\code{retrieved_number}}{the total number of observations}
\item{\code{retrieved_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe , whose variables include:
\code{term: 屆別},
\code{sessionPeriod: 會期},
\code{sessionTimes: 會次},
\code{meetingTimes: 臨時會會次},
\code{eyNumber: 行政院函公文編號},
\code{lyNumber:立法院函編號},
\code{subject: 案由},
\code{content: 內容},
\code{docUrl: 檔案下載位置},  and
\code{selectTerm: 屆別期別篩選條件}}
}
}
\description{
The Records of the Questions Answered by the Executives 公報質詢事項行政院答復資訊
}
\details{
\code{get_executive_response} produces a list, which contains \code{title},
\code{query_time}, \code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names},
\code{manual_info} and \code{data}. To retrieve the user manual and more information
about the data frame, please use \code{legisTaiwan::get_variable_info("get_executive_response")}.
質詢類: 提供公報質詢事項行政院答復資訊 (自第8屆第1會期起)。
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=2}
}
