% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\name{get_executive_response}
\alias{get_executive_response}
\title{The Records of Response to the Questions by the Executives 公報質詢事項行政院答復資訊}
\usage{
get_executive_response(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{integer, numeric or NULL. The default is NULL. The data is only
available from 8th term. 參數必須為數值。資料從自第8屆起，預設值為8。}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to NULL. 參數必須為數值。
\code{review_session_info()} generates each session period  available option period
in Minguo (Taiwan) calendar.}

\item{verbose}{logical, indicates whether \code{get_executive_response} should
print out detailed output when retrieving the data. The default is set to TRUE}
}
\value{
list contains: \describe{
\item{\code{title}}{the records of the questions answered by the executives}
\item{\code{query_time}}{the queried time}
\item{\code{retrieved_number}}{the total number of observations}
\item{\code{retrieved_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{term}}{屆別}
\item{\code{sessionPeriod}}{會期}
\item{\code{sessionTimes}}{會次}
\item{\code{meetingTimes}}{臨時會會次}
\item{\code{eyNumber}}{行政院函公文編號}
\item{\code{lyNumber}}{立法院函編號}
\item{\code{subject}}{案由}
\item{\code{content}}{內容}
\item{\code{docUrl}}{案由}
\item{\code{item}}{檔案下載位置}
\item{\code{item}}{檔案下載位置}
\item{\code{selectTerm}}{屆別期別篩選條件}
}
}
}
}
\description{
The Records of Response to the Questions by the Executives 公報質詢事項行政院答復資訊
}
\details{
\strong{\code{get_executive_response}} produces a list, which contains \code{title},
\code{query_time}, \code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names},
\code{manual_info} and \code{data}. To retrieve the user manual and more information, please
use \code{get_variable_info("get_executive_response")}.

#'@note To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_executive_response")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=2}.
質詢類: 提供公報質詢事項行政院答復資訊 (自第8屆第1會期起)。
}
\examples{
## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「行政院答復」
get_executive_response(term = 8, session_period = 1)

}
\seealso{
\code{get_variable_info("get_executive_response")}, \code{review_session_info()}
}
\author{
David Liao (davidycliao@gmail.com)
}
