% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_infos}
\alias{get_infos}
\title{Check function information on the website of Taiwan Legislative API}
\usage{
get_infos(x)
}
\arguments{
\item{x}{The parameter should be `get_parlquestions`, `get_legislators`,
`get_executive_response`, `get_bills`, or `get_meetings`}
}
\description{
Check function information on the website of Taiwan Legislative API
}
\examples{
get_infos("get_parlquestions")

get_infos("get_legislators")

get_infos("get_executive_response")

get_infos("get_bills")

get_infos("get_meetings")
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=6}
\url{https://data.ly.gov.tw/getds.action?id=16}
\url{https://data.ly.gov.tw/getds.action?id=2}
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=153}
}
