% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{get_legislators}
\alias{get_legislators}
\title{Retrieving legislator' demographic information and background}
\usage{
get_legislators(term = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric Must be a numeric vector.}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data. The default is TRUE.}
}
\value{
list contains: \describe{
\item{\code{query_time}}{the queried time}
\item{\code{queried_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe}
} contains query_time, queried_term, url ,variable_names, manual_info and data
}
\description{
Retrieving legislator' demographic information and background
}
\details{
\code{get_legislators} produces a list, which contains  \code{query_time},
\code{queried_term}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}.
}
\examples{
## To retrieve legislator' demographic from 8th
## 輸入「屆次」下載當屆立委資料
get_legislators(term = 8)


## To retrieve legislator' demographic from all of terms
## 輸入「空白」下載所有立委資料
get_legislators()
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=16}
}
