% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{get_legislators}
\alias{get_legislators}
\title{Retrieving legislator' demographic information and background}
\usage{
get_legislators(term = NULL, verbose = TRUE)
}
\arguments{
\item{term}{A number, which requests answered questions from the term. The parameter should be set in
a numeric vector.
retrieved in number, url, and computing time.}

\item{verbose}{The default is TRUE, which return the information of the return.
object.}
}
\value{
A list contains query_time, queried_term, url ,variable_names, manual_info and data
}
\description{
Retrieving legislator' demographic information and background
}
\examples{
## To retrieve legislator' demographic from 8th
## 輸入「屆次」下載當屆立委資料
get_legislators(term = 8)


## To retrieve legislator' demographic from all of terms
## 輸入「空白」下載所有立委資料
get_legislators()
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=16}
}
