% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{get_legislators}
\alias{get_legislators}
\title{Retrieving legislator' demographics via Taiwan Legislative Yuan API
歷屆委員資料}
\usage{
get_legislators(term = NULL, verbose = TRUE)
}
\arguments{
\item{term}{Requesting answered questions from the term. The parameter should be set in
a numeric format. The default value is 2.
retrieved in number, url, and computing time.}
}
\value{
A list object contains a tibble carrying term, name, ename,
sex, party, etc.
}
\description{
Retrieving legislator' demographics via Taiwan Legislative Yuan API
歷屆委員資料
}
\examples{
# to retrieve 8th therm
get_legislators(term = 8)
# to retrieve all of terms
get_legislators()
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=16}
}
