% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_meetings}
\alias{get_meetings}
\title{Retrieving the spoken meeting records 委員發言}
\usage{
get_meetings(
  start_date = NULL,
  end_date = NULL,
  meeting_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{numeric or double formatted by Taiwan calander. Requesting
meeting records starting from the date. If a double is used, it should specify
as Taiwan calendar format, e.g. 1090101.}

\item{end_date}{numeric or double formatted by Taiwan calander. Requesting
meeting records ending from the date. If a double is used, it should specify as Taiwan
calendar format, e.g. 1090110.}

\item{meeting_unit}{The default is NULL, which includes all meetings
between the starting date and the ending date.}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data.}
}
\value{
An object of the list, which contains query_time, retrieved_number,
meeting_unit, start_date_ad, end_date_ad, start_date, end_date, url,
variable_names, manual_info and data.
}
\description{
Retrieving the spoken meeting records 委員發言
}
\examples{
## query meeting records by a period of the dates in Taiwan ROC calender format
## 輸入「中華民國民年」下載「委員發言」
get_meetings(start_date = "1050120", end_date = "1050210")

## query meeting records by a period of the dates in Taiwan ROC calender format
## and a meeting
## 輸入「中華民國民年」與「審查會議或委員會名稱」下載會議審查資訊
get_meetings(start_date = 1060120, end_date = 1070310, meeting_unit = "內政委員會")

}
\seealso{
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
}
