% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_meetings}
\alias{get_meetings}
\title{The Spoken Meeting Records 委員發言}
\usage{
get_meetings(
  start_date = NULL,
  end_date = NULL,
  meeting_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{numeric Must be formatted in ROC Taiwan calendar, e.g. 1090101.}

\item{end_date}{numeric Must be formatted in ROC Taiwan calendar, e.g. 1090102.}

\item{meeting_unit}{NULL The default is NULL, which includes all meeting types
between the starting date and the ending date.}

\item{verbose}{logical, indicates whether \code{get_meetings} should print out
detailed output when retrieving the data.}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the spoken meeting records }
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of the observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual, \url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}; or use legisTaiwan::get_variable_info("get_meetings")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{smeeting_date}}{會議日期}
\item{\code{meeting_status}}{會議狀態}
\item{\code{meeting_name}}{會議名稱}
\item{\code{meeting_content}}{會議事由}
\item{\code{speechers}}{委員發言名單}
\item{\code{meeting_unit}}{主辦單位}
\item{\code{date_ad}}{西元年}
}
}
}
}
\description{
The Spoken Meeting Records 委員發言
}
\details{
\code{get_meetings} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}.To retrieve the user
manual and more information about the data frame, please use \code{legisTaiwan::get_variable_info("get_meetings")}.
Further Check Required: the user manuals seems to be inconsistent
with actual data. 資料似乎不一致，待確認。委員發言（取得最早時間不詳，待檢查。）
}
\seealso{
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
}
