% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_meetings}
\alias{get_meetings}
\title{The Spoken Meeting Records 委員發言}
\usage{
get_meetings(
  start_date = NULL,
  end_date = NULL,
  meeting_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{numeric Must be formatted in ROC Taiwan calendar, e.g. 1090101.}

\item{end_date}{numeric Must be formatted in ROC Taiwan calendar, e.g. 1090102.}

\item{meeting_unit}{NULL The default is NULL, which includes all meeting types
between the starting date and the ending date.}

\item{verbose}{logical, indicates whether \code{get_meetings} should print out
detailed output when retrieving the data.}
}
\value{
list, which contains: \describe{
\item{\code{title}}{Retrieving the spoken meeting records }
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number url of the page}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\code{smeeting_date: 會議日期},
\code{meeting_status: 會議狀態},
\code{meeting_name: 會議名稱},
\code{meeting_content: 會議事由},
\code{speechers: 委員發言名單},
\code{meeting_unit:主辦單位},
\code{date_ad: 西元年}}
}
}
\description{
The Spoken Meeting Records 委員發言
}
\details{
\code{get_meetings} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data}.To retrieve the user
manual and more information about the data frame, please use
\code{legisTaiwan::get_variable_info("get_meetings")}.
\url{+ Further Check Required: the user manuals seems to be inconsistent with actual data. 資料似乎不一致，待確認。}
委員發言（取得最早時間不詳，待檢查。）
}
\examples{
## query meeting records by a period of the dates in Taiwan ROC calender format
## 輸入「中華民國民年」下載「委員發言」
get_meetings(start_date = "1050120", end_date = "1050210")

## query meeting records by a period of the dates in Taiwan ROC calender format
## and a meeting
## 輸入「中華民國民年」與「審查會議或委員會名稱」下載會議審查資訊
get_meetings(start_date = 1060120, end_date = 1070310, meeting_unit = "內政委員會")

}
\seealso{
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
}
