% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_meetings}
\alias{get_meetings}
\title{Retrieving the spoken meeting records
下載「委員發言」}
\usage{
get_meetings(
  start_date = NULL,
  end_date = NULL,
  meeting_unit = NULL,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{Requesting meeting records starting from the date.
A double represents a date in ROC Taiwan format.
If a double is used, it should specify as Taiwan
calendar format, e.g. 1090110.}

\item{end_date}{Requesting meeting records ending from the date.
A double represents a date in ROC Taiwan format.
If a double is used, it should specify as Taiwan
calendar format, e.g. 1090110.}

\item{meeting_unit}{The default is NULL, which include all meetings
between the starting date and the ending date.}

\item{verbose}{The default value is TRUE, displaying the description
of data retrieved in number, url and computing time.}
}
\value{
A list carries a main tibble dataframe that contains the date, status,
name, content and speakers.
}
\description{
Retrieving the spoken meeting records
下載「委員發言」
}
\examples{
## query meeting records by a period of the dates in Taiwan ROC calender format
## 輸入「中華民國民年」下載「委員發言」
get_meetings(1050120, 1050210)

get_meetings("1050120", "1050210")

get_meetings(start_date = "1050120", end_date = "1050210")

## query meeting records by a period of the dates in Taiwan ROC calender format
## and a meeting
## 輸入「中華民國民年」與「審查會議或委員會名稱」下載會議審查資訊
get_meetings(start_date = 1060120, end_date = 1070310, meeting_unit = "內政委員會")


}
\seealso{
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
}
