% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\name{get_parlquestions}
\alias{get_parlquestions}
\title{Retrieving the records of parliamentary questions 委員質詢事項}
\usage{
get_parlquestions(term = 8, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{integer or numeric. The parameter should be set in a numeric
format. The default is 8. The data is only available from 8th
term 參數必須為數值，資料從立法院第8屆開始計算。}

\item{session_period}{integer or numeric. The session is between 1 and 8.
session_period 參數必須為數值。}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data. The default is TRUE}
}
\value{
list
}
\description{
Retrieving the records of parliamentary questions 委員質詢事項
}
\details{
Function produces a list, which contains query_time,
retrieved_number, retrieved_term, url, variable_names, manual_info and data.
}
\examples{
## presenting the description of parliamentary questions.
## 資料說明索引
get_variable_info("get_parlquestions")

## query parliamentary questions by term.
## 輸入「立委會期」下載立委質詢資料

get_parlquestions(term = 8)

## query parliamentary questions by term.
## 輸入「立委屆期」與「會期」下載立委質詢資料

get_parlquestions(term = 8, session_period = 2)


#' ## query parliamentary questions by term.
## 輸入「空白」下載立委全部質詢資料

get_parlquestions(term = 8, session_period = 2)
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=6}
}
