% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\name{get_parlquestions}
\alias{get_parlquestions}
\title{Retrieving the records of parliamentary questions
via Taiwan Legislative Yuan API 委員質詢事項}
\usage{
get_parlquestions(term = 8, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{Requesting questions from the term. The parameter should be set in
a numeric format. The default value is 8. The data is only available from 8th
term 參數必須為數值，資料從立法院第8屆開始計算。}

\item{session_period}{session in the term. The session is between 1 and 8.
session_period 參數必須為數值。}

\item{verbose}{The default value is TRUE, displaying the description of data
retrieved in number, url and computing time.}
}
\value{
A list object contains a tibble carrying
}
\description{
Retrieving the records of parliamentary questions
via Taiwan Legislative Yuan API 委員質詢事項
}
\examples{


## query parliamentary questions by term.
## 輸入「立委會期」下載立委質詢資料

get_parlquestions(term = 8)

## query parliamentary questions by term.
## 輸入「立委屆期」與「會期」下載立委質詢資料

get_parlquestions(term = 8, session_period = 2)


#' ## query parliamentary questions by term.
## 輸入「空白」下載立委全部質詢資料

get_parlquestions(term = 8, session_period = 2)
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=6}
}
