% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parlquestions.R
\name{get_parlquestions}
\alias{get_parlquestions}
\title{The Records of Parliamentary Questions 委員質詢事項資訊}
\usage{
get_parlquestions(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or NULL. The default is set to NULL. 參數必須為數值。}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to 8 參數必須為數值。
\code{review_session_info()} generates each session period available option period
in Minguo (Taiwan) calendar.}

\item{verbose}{logical, indicates whether \code{get_parlquestions} should print out
detailed output when retrieving the data. The default is TRUE.}
}
\value{
A list containing:
\describe{
\item{\code{title}}{the records of parliamentary questions}
\item{\code{query_time}}{the queried time}
\item{\code{retrieved_number}}{the total number of observations}
\item{\code{retrieved_term}}{the queried term}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual from \url{https://data.ly.gov.tw/getds.action?id=6}, or use get_variable_info("get_parlquestions")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{
\item{\code{term}}{屆別}
\item{\code{sessionPeriod}}{會期}
\item{\code{sessionTimes}}{會次}
\item{\code{item}}{項目}
\item{\code{selectTerm}}{屆別期別篩選條件}
}
}
}
}
\description{
The Records of Parliamentary Questions 委員質詢事項資訊
}
\details{
\code{get_parlquestions} produces a list, which contains \code{title},
\code{query_time}, \code{retrieved_number}, \code{retrieved_term}, \code{url}, \code{variable_names},
\code{manual_info}, and \code{data}.
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_parlquestions")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=6}.
質詢類: 提供議事日程本院委員之質詢事項資訊(自第8屆第1會期起)。
}
\examples{
## Query parliamentary questions by term.
## 輸入「立委會期」下載立委質詢資料
get_parlquestions(term = 10)

## Query parliamentary questions by term and session period.
## 輸入「立委屆期」與「會期」下載立委質詢資料
get_parlquestions(term = 10, session_period = 2)

}
\seealso{
\code{get_variable_info("get_parlquestions")}
}
\author{
David Liao (davidycliao@gmail.com)
}
