% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_public_debates}
\alias{get_public_debates}
\title{Retrieving the records of national public debates
議事類: 提供公報之國是論壇資訊，並包含書面意見。自第8屆第1會期起，
但實測資料從第10屆。}
\usage{
get_public_debates(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or NULL The default value is 10
參數必須為數值，資料從自第8屆第1會期起，但實測資料從第十屆。。}

\item{session_period}{integer or NULL. Available options for the session periods
is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is NULL. 參數必須為數值。}

\item{verbose}{logical, indicates whether \code{get_public_debates} should print out
detailed output when retrieving the data. The default is TRUE}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the offical manual}
\item{\code{data}}{a tibble dataframe, whose variables include:
\code{term: 屆別},
\code{sessionPeriod: 會期},
\code{sessionTimes: 會次},
\code{meetingTimes: 臨時會會次},
\code{dateTimeDesc:日期時間說明},
\code{meetingRoom: 會議地點},
\code{chairman:主持人},
\code{legislatorName: 委員名稱},
\code{speakType:發言類型(paper:書面發言,speak:發言)},
\code{content:內容}, and
\code{selectTerm:屆別期別篩選條件}}
}
}
\description{
Retrieving the records of national public debates
議事類: 提供公報之國是論壇資訊，並包含書面意見。自第8屆第1會期起，
但實測資料從第10屆。
}
\examples{
## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載國是論壇資訊。
get_public_debates(term = 10, session_period = 2)

}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=7}
}
