% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_public_debates}
\alias{get_public_debates}
\title{Retrieving the records of national public debates 下載「國是論壇」資料}
\usage{
get_public_debates(term = 8, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{Requesting answered questions by term. The parameter should be set
in a numeric vector The default value is 8. The data is only available from
the 8th term 參數必須為數值，資料從立法院「第8屆」開始計算。}

\item{session_period}{legislative session in the term. The session is between
1 and 8. 參數必須為數值。The parameter should be set in a numeric vector.}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data. The default value is TRUE}
}
\value{
A list contains query_time, retrieved_number, start_date_ad,
end_date_ad, start_date, end_date, url, variable_names, manual_info and data
}
\description{
Retrieving the records of national public debates 下載「國是論壇」資料
}
\examples{
## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「質詢事項 (行政院答復部分)」
get_public_debates(term = 10, session_period = 2)

get_public_debates(term = 10, session_period = 1)

}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=7}
}
