% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_public_debates}
\alias{get_public_debates}
\title{The Records of National Public Debates 國是論壇}
\usage{
get_public_debates(term = NULL, session_period = NULL, verbose = TRUE)
}
\arguments{
\item{term}{numeric or NULL The default is set to 10. 參數必須為數值，資料從自
第8屆第1會期起，但實測資料從第10屆，故預設為10。}

\item{session_period}{integer, numeric or NULL. Available
options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to NULL. 參數必須為數值。
\code{review_session_info()} generates each session period  available option period
in Minguo (Taiwan) calendar.}

\item{verbose}{logical, indicates whether \code{get_public_debates} should print out
detailed output when retrieving the data. The default is TRUE}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the official manual, \url{https://data.ly.gov.tw/getds.action?id=7}; or use get_variable_info("get_public_debates")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{term}}{屆別}
\item{\code{sessionPeriod}}{會期}
\item{\code{sessionTimes}}{會次}
\item{\code{meetingTimes}}{臨時會會次}
\item{\code{dateTimeDesc}}{日期時間說明}
\item{\code{meetingRoom}}{會議地點}
\item{\code{chairman}}{主持人}
\item{\code{legislatorName}}{委員姓名}
\item{\code{speakType}}{發言類型(paper:書面發言,speak:發言)}
\item{\code{content}}{內容}
\item{\code{selectTerm}}{屆別期別篩選條件}
}
}
}
}
\description{
The Records of National Public Debates 國是論壇
}
\details{
\code{get_public_debates} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data.}
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_public_debates")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=7}.
議事類: 提供公報之國是論壇資訊，並包含書面意見。自第8屆第1會期起，但實測資料從第10屆。
}
\examples{
## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載國是論壇資訊。
get_public_debates(term = 10, session_period = 2)

}
\seealso{
\code{get_variable_info("get_public_debates")}, \code{review_session_info()}

Regarding Minguo calendar, please see \url{https://en.wikipedia.org/wiki/Republic_of_China_calendar}.
}
