% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting.R
\name{get_speech_video}
\alias{get_speech_video}
\title{The Video Information of Meetings and Committees 院會及委員會之委員發言片段相關影片資訊}
\usage{
get_speech_video(start_date = NULL, end_date = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{character Must be formatted in Minguo (ROC) calendar with three
forward slashes between year, month and day, e.g. "106/10/20".}

\item{end_date}{character Must be formatted in Minguo (ROC) calendar  with three
forward slashes between year, month and day, e.g. "109/01/10".}

\item{verbose}{logical, indicates whether get_meetings should print out
detailed output when retrieving the data.}
}
\value{
list, which contains: \describe{
\item{\code{title}}{the meeting records of cross-caucus session}
\item{\code{query_time}}{the query time}
\item{\code{retrieved_number}}{the number of observation}
\item{\code{meeting_unit}}{the meeting unit}
\item{\code{start_date_ad}}{the start date  in POSIXct}
\item{\code{end_date_ad}}{the end date in POSIXct}
\item{\code{start_date}}{the start date in ROC Taiwan calendar}
\item{\code{url}}{the retrieved json url}
\item{\code{variable_names}}{the variables of the tibble dataframe}
\item{\code{manual_info}}{the official manual, \url{https://data.ly.gov.tw/getds.action?id=148}; or use get_variable_info("get_speech_video")}
\item{\code{data}}{a tibble dataframe, whose variables include:
\describe{\item{\code{term}}{屆別}
\item{\code{sessionPeriod}}{會期}
\item{\code{meetingDate}}{會議日期(西元年)}
\item{\code{meetingTime}}{會議時間}
\item{\code{meetingTypeName}}{主辦單位}
\item{\code{meetingName}}{會議名稱}
\item{\code{meetingContent}}{會議事由}
\item{\code{legislatorName}}{委員姓名}
\item{\code{areaName}}{選區名稱}
\item{\code{speechStartTime}}{委員發言時間起}
\item{\code{speechEndTime}}{委員發言時間迄}
\item{\code{speechRecordUrl}}{發言紀錄網址}
\item{\code{videoLength}}{影片長度}
\item{\code{videoUrl}}{影片網址}
\item{\code{selectTerm}}{屆別期別篩選條件}
}
}
}
}
\description{
The Video Information of Meetings and Committees 院會及委員會之委員發言片段相關影片資訊
}
\details{
\code{get_speech_video} produces a list, which contains \code{title}, \code{query_time},
\code{retrieved_number}, \code{meeting_unit}, \code{start_date_ad}, \code{end_date_ad}, \code{start_date},
\code{end_date}, \code{url}, \code{variable_names}, \code{manual_info} and \code{data.} To retrieve the user
manual and more information about the data frame, please use \code{get_variable_info("get_speech_video")}.
}
\note{
To retrieve the user manual and more information about variable of the data
frame, please use \code{get_variable_info("get_speech_video")}
or visit the API manual at \url{https://data.ly.gov.tw/getds.action?id=148}.
會議類:提供立法院院會及委員會之委員發言片段相關影片資訊 (自第9屆第1會期起)。
}
\examples{
## query full video information of meetings and committees using a period of
## the dates in Taiwan ROC calender format with forward slash (/).
## 輸入「中華民國民年」下載「委員發言片段相關影片資訊」，輸入時間請依照該
## 格式 "105/10/20"，需有「正斜線」做隔開。
get_speech_video(start_date = "105/10/20", end_date = "109/03/10")

}
\seealso{
\code{get_variable_info("get_speech_video")}
}
