% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_variabel_infos}
\alias{get_variabel_infos}
\title{Check all function information}
\usage{
get_variabel_infos(param_)
}
\arguments{
\item{param_}{characters. The parameter should be \code{get_parlquestions},
\code{get_legislators}, \code{get_executive_response}, \code{get_bills}, \code{get_meetings},
\code{get_caucus_meetings} or \code{get_public_debates},}
}
\description{
Check all function information
}
\seealso{
\url{https://data.ly.gov.tw/getds.action?id=6}
\url{https://data.ly.gov.tw/getds.action?id=16}
\url{https://data.ly.gov.tw/getds.action?id=2}
\url{https://data.ly.gov.tw/getds.action?id=8}
\url{https://data.ly.gov.tw/getds.action?id=148}
\url{https://data.ly.gov.tw/getds.action?id=1}
\url{https://data.ly.gov.tw/getds.action?id=7}
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}
\url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=153}
}
