% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_variable_info}
\alias{get_variable_info}
\title{Check Each Function's Manual}
\usage{
get_variable_info(param_)
}
\arguments{
\item{param_}{characters. Must be one of options below: \describe{
\item{get_bills}{get_bills: the records of the bills, see \url{https://data.ly.gov.tw/getds.action?id=6}}
\item{get_bills_2}{the records of legislators and the government proposals, see \url{https://data.ly.gov.tw/getds.action?id=6}}
\item{get_meetings}{the spoken meeting records, see \url{https://www.ly.gov.tw/Pages/List.aspx?nodeid=154}}
\item{get_caucus_meetings}{the meeting records of cross-caucus session, see \url{https://data.ly.gov.tw/getds.action?id=8}}
\item{get_speech_video}{the full video information of meetings and committees, see \url{https://data.ly.gov.tw/getds.action?id=148}}
\item{get_public_debates}{the records of national public debates, see \url{https://data.ly.gov.tw/getds.action?id=7}}
\item{get_parlquestions}{the records of parliamentary questions, see \url{https://data.ly.gov.tw/getds.action?id=6}}
\item{get_executive_response}{the records of the questions answered by the executives, see \url{https://data.ly.gov.tw/getds.action?id=2}}
}}
}
\value{
list \describe{
\item{\code{page_info}}{information of the end point}
\item{\code{reference_url}}{the url of the page}}
}
\description{
\code{get_variable_info} generate each API's endpoint manual returned
from the website of Taiwan Legislative Yuan. The avalaible options is: \code{get_bills},
\code{get_bills_2}, \code{get_meetings}, \code{get_caucus_meetings}, \code{get_speech_video} ,
\code{get_public_debates}, \code{get_parlquestions}, \code{get_executive_response} and
\code{get_committee_record}.
}
\details{
\code{get_variable_info} produces a list, which contains \code{page_info} and \code{reference_url}.
}
\examples{
\dontrun{
get_variable_info("get_bills")
}
}
\seealso{
\code{review_session_info()}.
}
\author{
David Liao (davidycliao@gmail.com)
}
