% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{review_session_info}
\alias{review_session_info}
\title{Check Session Periods in Each ROC Year}
\usage{
review_session_info(term_, ...)
}
\arguments{
\item{term_}{characters or numeric. M}
}
\value{
dataframe
}
\description{
Check Session Periods in Each ROC Year
}
\details{
\code{review_session_info} produces a dataframe, displaying each session
period in year formated in Minguo (ROC) calendar.
}
