% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformed_date_bill}
\alias{transformed_date_bill}
\title{Transforming bill proposed date in Taiwan ROC calendar to A.D. format}
\usage{
transformed_date_bill(ROCdate)
}
\arguments{
\item{rocdate}{date format in Taiwan ROC calendar (e.g., "1050531") as a string vector}
}
\value{
date format in A.D. format
}
\description{
Transforming bill proposed date in Taiwan ROC calendar to A.D. format
}
\examples{
transformed_date_bill("1050531")
}
