% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformed_date_bill}
\alias{transformed_date_bill}
\title{Transforming Minguo (Taiwan) Calendar to A.D. Calendar II}
\usage{
transformed_date_bill(roc_date)
}
\arguments{
\item{roc_date}{date format in Taiwan ROC calendar (e.g., "1050531") in a character vector}
}
\value{
date in POSIXct
}
\description{
\code{transformed_date_meeting()} transforms Minguo (Taiwan) Calendar
to A.D. format in POSIXct for \code{get_bill()}, e.g. "1050531" to "2016-05-31".
}
\examples{
\dontrun{
transformed_date_bill("1050531")
}
}
\keyword{internal}
