% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformed_date_meeting}
\alias{transformed_date_meeting}
\title{Transforming Minguo (Taiwan) Calendar to A.D. Calendar I}
\usage{
transformed_date_meeting(roc_date)
}
\arguments{
\item{roc_date}{Date format in Minguo (Taiwan) calendar (e.g., "105/05/31") as a
string vector}
}
\value{
date in POSIXct
}
\description{
\code{transformed_date_meeting()} transforms Minguo (Taiwan) Calendar
to A.D. calendar in POSIXct for \code{get_meetings()}, \code{get_caucus_meetings()},
and \code{get_speech_video()},
}
\details{
\code{check_date} transforms ROC date to a date in POSIXct, e.g. "105/05/31" to "2016-05-31".
}
\examples{
\dontrun{
transformed_date_meeting("105/05/31")
}

}
\keyword{internal}
