% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{website_availability}
\alias{website_availability}
\title{A Check for the Website Availability I}
\usage{
website_availability(site = "https://data.ly.gov.tw/index.action")
}
\arguments{
\item{site}{https://data.ly.gov.tw/index.action}
}
\description{
A Check for the Website Availability I
}
\seealso{
\code{check_internet()},
}
