% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{website_availability}
\alias{website_availability}
\title{A check for the website availability and connection}
\usage{
website_availability(site = "https://data.ly.gov.tw/index.action")
}
\arguments{
\item{site}{https://data.ly.gov.tw/index.action}
}
\description{
A check for the website availability and connection
}
\seealso{
\url{https://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r?noredirect=1&lq=1}
}
