% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{website_availability}
\alias{website_availability}
\title{Check for the Website Availability I}
\usage{
website_availability(site = "https://data.ly.gov.tw/index.action")
}
\arguments{
\item{site}{A website URL to check. Default is "https://data.ly.gov.tw/index.action".}
}
\description{
This function checks the availability of a specified website by trying to read
the first line of the site's content.
}
\seealso{
\code{check_internet()}
}
\keyword{internal}
