% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{website_availability2}
\alias{website_availability2}
\title{Check for the Website Availability II}
\usage{
website_availability2(
  site = "https://npl.ly.gov.tw/do/www/appDate?status=0&expire=02&startYear=0"
)
}
\arguments{
\item{site}{A website URL to check. Default is "https://npl.ly.gov.tw/do/www/appDate?status=0&expire=02&startYear=0".}
}
\description{
This function checks the availability of a specified website by trying to read
the first line of the site's content.
}
\seealso{
\code{check_internet()} and \code{website_availability()}.
}
\keyword{internal}
